+++ test/median_test.rb
@@ -70,7 +70,7 @@ class MedianTest < Minitest::Test
     assert_equal "A", User.group(:name).order(:name).average(:visits_count).keys.first
     assert_equal "B", User.group(:name).order("average_visits_count desc").average(:visits_count).keys.first
     assert_equal "A", User.group(:name).order(:name).median(:visits_count).keys.first
-    # assert_equal "B", User.group(:name).order("median_visits_count desc").median(:visits_count).keys.first
+    assert_equal "B", User.group(:name).order("median_visits_count desc").median(:visits_count).keys.first
   end
 
   def test_expression